fix(build.bash): specific signed type for char
In most cases, char is unsigned on ARM (for performance reasons) and signed on other platforms.
https://stackoverflow.com/questions/20576300/is-char-signed-or-unsigned-by-default-on-ios

So if our system/toolchain defaults to an unsigned char type. That means when getopt() returns -1,
it gets converted to 255 and stored in t.
https://stackoverflow.com/questions/17070958/c-why-does-getopt-return-255-on-linux

The workaround is specific "char" type to signed by default for SAC is
developed and delivered mostly in Intel-x86 environment (that assumes
char is signed)